fix properties in entites, and change double for decimal

+++ Entities/Game.cs
@@ -6,14 +6,14 @@ public class Game
     public string Name { get; set; } = String.Empty;
     public string Summary { get; set; } = String.Empty;
     public DateTime Date { get; set; }
-    public double Score { get; set; } = 0.0;
+    public decimal Score { get; set; } = 0.0m;
     public string Director { get; set; } = String.Empty;
     public string? Writers { get; set; }
     public string? Company { get; set; }
     
     public Game() { }
 
-    public Game(string name, string summary, DateTime date, double score, string director, string? writers, string? company)
+    public Game(string name, string summary, DateTime date, decimal score, string director, string? writers, string? company)
     {
         Name = name;
         Summary = summary;
